tests/wpt/web-platform-tests/preload/single_download_late_used_preload.html

<!DOCTYPE html>
<title>Ensure preloaded resources are not downloaded again when used</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel=preload href="resources/square.png?pipe=trickle(d1)" as=image>
<script>
    var link = document.getElementsByTagName("link")[0]
    assert_equals(link.as, "image");
    link.addEventListener("load", () => {
        assert_equals(performance.getEntriesByType("resource").length, 3)
        var img = document.createElement("img");
        img.src = "resources/square.png?pipe=trickle(d1)";
        img.onload = () => {
            assert_equals(performance.getEntriesByType("resource").length, 3);
            done();
        };
        document.body.appendChild(img);
    });
</script>
<body>
